attempt to fix cypress CI
