+++ bindings/wasm/identity_wasm/examples/src/util.ts
@@ -12,17 +12,14 @@ import {
     Storage,
     StorageSigner,
     Transaction,
-    TransactionSigner,
 } from "@iota/identity-wasm/node";
 import { CoreClientReadOnly } from "@iota/iota-interaction-ts/node/core_client";
 import { getFullnodeUrl, IotaClient, TransactionEffects } from "@iota/iota-sdk/client";
 import { getFaucetHost, requestIotaFromFaucetV0 } from "@iota/iota-sdk/faucet";
 import { IotaEvent } from "@iota/iota-sdk/src/client/types/generated";
 import { Transaction as SdkTransaction } from "@iota/iota-sdk/transactions";
-import { NotarizationClient, NotarizationClientReadOnly } from "@iota/notarization/node";
 
 export const IOTA_IDENTITY_PKG_ID = globalThis?.process?.env?.IOTA_IDENTITY_PKG_ID || "";
-export const IOTA_NOTARIZATION_PKG_ID = globalThis?.process?.env?.IOTA_NOTARIZATION_PKG_ID || "";
 export const NETWORK_NAME_FAUCET = globalThis?.process?.env?.NETWORK_NAME_FAUCET || "localnet";
 export const NETWORK_URL = getFullnodeUrl(NETWORK_NAME_FAUCET);
 
@@ -114,18 +111,4 @@ export class SendZeroCoinTx implements Transaction<string> {
     ): Promise<string> {
         return await this.apply(effects, client);
     }
-}
-
-export async function getNotarizationClient(signer: TransactionSigner): Promise<NotarizationClient> {
-    if (!IOTA_NOTARIZATION_PKG_ID) {
-        throw new Error(`IOTA_NOTARIZATION_PKG_ID env variable must be provided to run the notarization examples`);
-    }
-
-    const iotaClient = new IotaClient({ url: NETWORK_URL });
-    const notarizationClientReadOnly = await NotarizationClientReadOnly.createWithPkgId(
-        iotaClient,
-        IOTA_NOTARIZATION_PKG_ID,
-    );
-
-    return await NotarizationClient.create(notarizationClientReadOnly, signer);
 }
